restructure files, refactor
