+++ src/components/Explorer/QueryBuilder/d3RangeSlider.js
