+++ src/components/Explorer/Result/Chart/ChartHelpers.jsx
