+++ src/components/Explorer/Result/Chart/RunChart.jsx
@@ -17,7 +17,7 @@ import {
     drawLegend,
     addHeaders,
     addColumns
-} from './ResultChartHelpers';
+} from './ChartHelpers';
 
 const chartId = "runResultChart"
